Update Contracts, Test & App
Modify uint256 to uint to propperly have a clearer code in contracts CustomCrowdsale.sol & CustomToken.sol
Add new requirement coded as a modifier "whenICONotCompleted" to the function buyToken() in CustomCrowdsale.sol so it can prevent the buy after the closeTime, launching a revert due to time constraints.

Update Test to properly reflect the values during the console outputs. 
TODO: Take the Size of the string and apply slice to properly change the current value and not use the hardcoded one. 

Update App Components to properly reflect the values being exchanged. 
TODO: apply the same fix to the values being displayed in the App.

ClaimToken function in App.js properly transfer 2e18 instead of 2e-18.